[Http] Reorganize factory classes in Message namespace.
