+++ src/Valkyrja/Http/Message/Provider/ServiceProvider.php
@@ -16,8 +16,8 @@ namespace Valkyrja\Http\Message\Provider;
 use Override;
 use Valkyrja\Container\Manager\Contract\ContainerContract;
 use Valkyrja\Container\Provider\Provider;
-use Valkyrja\Http\Message\Factory\Contract\ResponseFactoryContract;
-use Valkyrja\Http\Message\Factory\ResponseFactory;
+use Valkyrja\Http\Message\Response\Factory\Contract\ResponseFactoryContract;
+use Valkyrja\Http\Message\Response\Factory\ResponseFactory;
 
 final class ServiceProvider extends Provider
 {
