+++ src/Valkyrja/Http/Message/Request/Factory/ServerFactory.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Http\Message\Factory;
+namespace Valkyrja\Http\Message\Request\Factory;
 
 use function function_exists;
