+++ src/Valkyrja/Http/Message/Request/Psr/Request.php
@@ -19,10 +19,10 @@ use Psr\Http\Message\StreamInterface;
 use Psr\Http\Message\UriInterface;
 use Valkyrja\Http\Message\Enum\ProtocolVersion;
 use Valkyrja\Http\Message\Enum\RequestMethod;
-use Valkyrja\Http\Message\Factory\HeaderFactory;
-use Valkyrja\Http\Message\Factory\StreamFactory;
+use Valkyrja\Http\Message\Header\Factory\HeaderFactory;
 use Valkyrja\Http\Message\Header\Header;
 use Valkyrja\Http\Message\Request\Contract\RequestContract;
+use Valkyrja\Http\Message\Stream\Factory\StreamFactory;
 use Valkyrja\Http\Message\Stream\Psr\Stream;
 use Valkyrja\Http\Message\Uri\Psr\Uri;
 use Valkyrja\Http\Message\Uri\Uri as ValkyrjaUri;
