+++ src/Valkyrja/Http/Message/Request/Psr/ServerRequest.php
@@ -16,8 +16,8 @@ namespace Valkyrja\Http\Message\Request\Psr;
 use Override;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\UploadedFileInterface;
-use Valkyrja\Http\Message\Factory\UploadedFileFactory;
 use Valkyrja\Http\Message\File\Contract\UploadedFileContract;
+use Valkyrja\Http\Message\File\Factory\UploadedFileFactory;
 use Valkyrja\Http\Message\File\Psr\UploadedFile;
 use Valkyrja\Http\Message\Request\Contract\ServerRequestContract;
 use Valkyrja\Http\Message\Request\Throwable\Exception\RuntimeException;
