+++ src/Valkyrja/Http/Message/Response/Factory/Contract/ResponseFactoryContract.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Http\Message\Factory\Contract;
+namespace Valkyrja\Http\Message\Response\Factory\Contract;
 
 use Valkyrja\Http\Message\Enum\StatusCode;
 use Valkyrja\Http\Message\Header\Contract\HeaderContract;
