+++ src/Valkyrja/Http/Message/Uri/Factory/UriFactory.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Http\Message\Factory;
+namespace Valkyrja\Http\Message\Uri\Factory;
 
 use Psr\Http\Message\UriInterface;
 use Valkyrja\Http\Message\Header\Contract\HeaderContract;
