+++ src/Valkyrja/Http/Routing/Controller/Controller.php
@@ -13,8 +13,8 @@ declare(strict_types=1);
 
 namespace Valkyrja\Http\Routing\Controller;
 
-use Valkyrja\Http\Message\Factory\Contract\ResponseFactoryContract;
 use Valkyrja\Http\Message\Request\Contract\ServerRequestContract;
+use Valkyrja\Http\Message\Response\Factory\Contract\ResponseFactoryContract;
 
 abstract class Controller
 {
