+++ src/Valkyrja/Http/Routing/Factory/ResponseFactory.php
@@ -15,8 +15,8 @@ namespace Valkyrja\Http\Routing\Factory;
 
 use Override;
 use Valkyrja\Http\Message\Enum\StatusCode;
-use Valkyrja\Http\Message\Factory\Contract\ResponseFactoryContract as HttpMessageResponseFactory;
 use Valkyrja\Http\Message\Response\Contract\RedirectResponseContract;
+use Valkyrja\Http\Message\Response\Factory\Contract\ResponseFactoryContract as HttpMessageResponseFactory;
 use Valkyrja\Http\Routing\Factory\Contract\ResponseFactoryContract as Contract;
 use Valkyrja\Http\Routing\Url\Contract\UrlContract;
