+++ src/Valkyrja/Http/Server/Psr/RequestHandler.php
@@ -17,7 +17,7 @@ use Override;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\RequestHandlerInterface;
-use Valkyrja\Http\Message\Factory\RequestFactory;
+use Valkyrja\Http\Message\Request\Factory\RequestFactory;
 use Valkyrja\Http\Message\Response\Psr\Response;
 use Valkyrja\Http\Server\Handler\Contract\RequestHandlerContract;
