+++ src/Valkyrja/View/Factory/ResponseFactory.php
@@ -15,9 +15,9 @@ namespace Valkyrja\View\Factory;
 
 use Override;
 use Valkyrja\Http\Message\Enum\StatusCode;
-use Valkyrja\Http\Message\Factory\Contract\ResponseFactoryContract as HttpMessageResponseFactoryContract;
-use Valkyrja\Http\Message\Factory\ResponseFactory as HttpMessageResponseFactory;
 use Valkyrja\Http\Message\Response\Contract\ResponseContract;
+use Valkyrja\Http\Message\Response\Factory\Contract\ResponseFactoryContract as HttpMessageResponseFactoryContract;
+use Valkyrja\Http\Message\Response\Factory\ResponseFactory as HttpMessageResponseFactory;
 use Valkyrja\View\Factory\Contract\ResponseFactoryContract as Contract;
 use Valkyrja\View\Renderer\Contract\RendererContract;
 use Valkyrja\View\Renderer\PhpRenderer;
