+++ tests/Functional/Abstract/TestCase.php
@@ -16,8 +16,8 @@ namespace Valkyrja\Tests\Functional\Abstract;
 use PHPUnit\Framework\TestCase as PHPUnitTestCase;
 use Valkyrja\Application\Entry\Abstract\App;
 use Valkyrja\Application\Kernel\Contract\ApplicationContract;
-use Valkyrja\Http\Message\Factory\RequestFactory;
 use Valkyrja\Http\Message\Request\Contract\ServerRequestContract;
+use Valkyrja\Http\Message\Request\Factory\RequestFactory;
 use Valkyrja\Support\Directory\Directory;
 use Valkyrja\Tests\EnvClass;
