+++ tests/Unit/Api/Manager/ApiTest.php
@@ -21,8 +21,8 @@ use Valkyrja\Api\Manager\Api;
 use Valkyrja\Api\Manager\Contract\ApiContract;
 use Valkyrja\Api\Model\Contract\JsonContract;
 use Valkyrja\Http\Message\Enum\StatusCode;
-use Valkyrja\Http\Message\Factory\Contract\ResponseFactoryContract;
 use Valkyrja\Http\Message\Response\Contract\JsonResponseContract;
+use Valkyrja\Http\Message\Response\Factory\Contract\ResponseFactoryContract;
 use Valkyrja\Http\Message\Throwable\Exception\HttpException;
 use Valkyrja\Orm\Entity\Contract\EntityContract;
 use Valkyrja\Tests\Unit\Abstract\TestCase;
