+++ tests/Unit/Api/Middleware/ApiThrowableCaughtMiddlewareTest.php
@@ -20,10 +20,10 @@ use Valkyrja\Api\Manager\Contract\ApiContract;
 use Valkyrja\Api\Middleware\ApiThrowableCaughtMiddleware;
 use Valkyrja\Api\Model\Json;
 use Valkyrja\Http\Message\Enum\StatusCode;
-use Valkyrja\Http\Message\Factory\Contract\ResponseFactoryContract;
 use Valkyrja\Http\Message\Request\Contract\ServerRequestContract;
 use Valkyrja\Http\Message\Response\Contract\JsonResponseContract;
 use Valkyrja\Http\Message\Response\Contract\ResponseContract;
+use Valkyrja\Http\Message\Response\Factory\Contract\ResponseFactoryContract;
 use Valkyrja\Http\Middleware\Contract\ThrowableCaughtMiddlewareContract;
 use Valkyrja\Http\Middleware\Handler\Contract\ThrowableCaughtHandlerContract;
 use Valkyrja\Tests\Unit\Abstract\TestCase;
