+++ tests/Unit/Http/Routing/Controller/ApiControllerTest.php
@@ -18,8 +18,8 @@ use Override;
 use Valkyrja\Api\Constant\Status;
 use Valkyrja\Api\Manager\Api;
 use Valkyrja\Http\Message\Enum\StatusCode;
-use Valkyrja\Http\Message\Factory\ResponseFactory;
 use Valkyrja\Http\Message\Request\ServerRequest;
+use Valkyrja\Http\Message\Response\Factory\ResponseFactory;
 use Valkyrja\Http\Message\Throwable\Exception\HttpException;
 use Valkyrja\Tests\Classes\Http\Routing\Controller\ApiControllerClass;
 use Valkyrja\Tests\Unit\Abstract\TestCase;
