+++ tests/Unit/Http/Routing/Factory/ResponseFactoryTest.php
@@ -17,8 +17,8 @@ use Override;
 use Valkyrja\Dispatch\Data\MethodDispatch;
 use Valkyrja\Http\Message\Constant\HeaderName;
 use Valkyrja\Http\Message\Enum\StatusCode;
-use Valkyrja\Http\Message\Factory\ResponseFactory as MessageResponseFactory;
 use Valkyrja\Http\Message\Header\Header;
+use Valkyrja\Http\Message\Response\Factory\ResponseFactory as MessageResponseFactory;
 use Valkyrja\Http\Routing\Collection\Collection;
 use Valkyrja\Http\Routing\Data\Route;
 use Valkyrja\Http\Routing\Factory\ResponseFactory;
